# Conflicts:
#	pkg/executor/adapter.go
#	pkg/executor/adapter_test.go
#	pkg/server/internal/resultset/resultset.go

+++ pkg/executor/adapter_test.go
@@ -491,7 +491,7 @@ func TestFinishExecuteStmtReportsTiDBRUV2WithoutSyncingRUDetails(t *testing.T) {
 	require.Equal(t, float64(23456), ruDetails.TiKVRUV2())
 	require.Equal(t, "rg1", reporter.group)
 	require.Equal(t, float64(23456), reporter.tikvRUV2)
-	require.Equal(t, float64(expected), reporter.tidbRUV2)
+	require.Equal(t, expected, reporter.tidbRUV2)
 	require.Equal(t, float64(412), reporter.tiflashRU)
 }
